requirements.txt
Flask==2.3.3
python-docx==0.8.11
striprtf==0.0.24
Werkzeug==2.3.7